volume_controller/requirements.txt
absl-py==2.0.0
attrs==23.2.0
cffi==1.16.0
comtypes==1.2.1
contourpy==1.2.0
cycler==0.12.1
flatbuffers==23.5.26
fonttools==4.47.2
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.8.2
mediapipe==0.10.9
numpy==1.26.3
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
protobuf==3.20.3
psutil==5.9.7
pyalsaaudio==0.10.0
pycaw==20230407
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
sounddevice==0.4.6
zipp==3.17.0
